:star: refactor files
