+++ index.html
@@ -370,7 +370,7 @@
 	<a class="gotopbtn clr-wt" id="scroll-top" href="#"> <i class="fa fa-arrow-up clr-wt"></i> </a>
 
 </body>
-<script src="./lib/particle.js"></script>
+<script src="./scripts/particle.js"></script>
 <script src="./scripts/particle-js-effect.js"></script>
 
 </html>
