+++ team.html
@@ -32,7 +32,6 @@
 	<link rel="stylesheet" href="css/style.css" type="text/css">
 	<link rel="stylesheet" href="css/team.css" type="text/css">
 	<!-- Defined scripts -->
-	<script src="./scripts/team.js"></script>
 	<script src="./scripts/app.js"></script>
 </head>
